perf(cli): defer push/init/locations deps off the run path
`cli.ts` statically imported the entire push + generator subsystem, so
every invocation — including `run` (the Heartbeat hot path) — eagerly
loaded archiver, undici, yaml, semver, unzip-stream and enquirer even
though only `push`/`init`/`locations` use them.

- Import push/generator/locations lazily inside their command actions.
- Lazy-load undici inside setGlobalProxy (only push/locations call it);
  the function is now async and its two callers await it.
- Move the pure parseSpaces helper from push/monitor to dsl/monitor so
  core/runner can use it without dragging archiver/yaml onto the run
  path. Re-exported from push/monitor for backward compatibility.

Measured on a `--dry-run`: require.cache entries drop from 519 to 150;
archiver, undici, yaml, unzip-stream, enquirer, lodash, compress-commons
and crc32-stream no longer load for a run.
